refactor(extension-loader): extract loading bundle extension logic
